extract NavbarNavLink
